requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via crawl4ai
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   litellm
    #   realtime
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.20.0
    # via crawl4ai
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anthropic==0.45.2
    # via pydantic-ai-slim
anyio==4.8.0
    # via
    #   anthropic
    #   groq
    #   httpx
    #   openai
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via crawl4ai
blinker==1.9.0
    # via streamlit
cachetools==5.5.1
    # via
    #   google-auth
    #   streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   litellm
    #   nltk
    #   streamlit
cohere==5.13.11
    # via pydantic-ai-slim
colorama==0.4.6
    # via
    #   click
    #   crawl4ai
    #   griffe
    #   tqdm
crawl4ai==0.4.248
    # via rag-owu (pyproject.toml)
cryptography==44.0.0
    # via pyopenssl
cssselect==1.2.0
    # via crawl4ai
deprecation==2.1.0
    # via postgrest
distro==1.9.0
    # via
    #   anthropic
    #   groq
    #   openai
eval-type-backport==0.2.2
    # via
    #   mistralai
    #   pydantic-ai-slim
fake-http-header==0.3.5
    # via tf-playwright-stealth
fake-useragent==2.0.3
    # via crawl4ai
fastavro==1.10.0
    # via cohere
filelock==3.17.0
    # via huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-auth==2.38.0
    # via pydantic-ai-slim
gotrue==2.11.3
    # via supabase
greenlet==3.1.1
    # via playwright
griffe==1.5.6
    # via pydantic-ai-slim
groq==0.16.0
    # via pydantic-ai-slim
h11==0.14.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via
    #   anthropic
    #   cohere
    #   crawl4ai
    #   gotrue
    #   groq
    #   litellm
    #   mistralai
    #   ollama
    #   openai
    #   postgrest
    #   pydantic-ai-slim
    #   pydantic-graph
    #   storage3
    #   supabase
    #   supafunc
httpx-sse==0.4.0
    # via cohere
huggingface-hub==0.28.1
    # via tokenizers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via litellm
jinja2==3.1.5
    # via
    #   altair
    #   litellm
    #   pydeck
jiter==0.8.2
    # via
    #   anthropic
    #   openai
joblib==1.4.2
    # via nltk
jsonpath-python==1.0.6
    # via mistralai
jsonschema==4.23.0
    # via
    #   altair
    #   litellm
jsonschema-specifications==2024.10.1
    # via jsonschema
litellm==1.60.0
    # via crawl4ai
logfire-api==3.4.0
    # via
    #   pydantic-ai-slim
    #   pydantic-graph
lxml==5.3.0
    # via crawl4ai
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mistralai==1.5.0
    # via pydantic-ai-slim
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.24.2
    # via altair
nltk==3.9.1
    # via crawl4ai
numpy==2.2.2
    # via
    #   crawl4ai
    #   pandas
    #   pydeck
    #   rank-bm25
    #   streamlit
ollama==0.4.7
    # via rag-owu (pyproject.toml)
openai==1.61.0
    # via
    #   rag-owu (pyproject.toml)
    #   litellm
    #   pydantic-ai-slim
packaging==24.2
    # via
    #   altair
    #   deprecation
    #   huggingface-hub
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==10.4.0
    # via
    #   crawl4ai
    #   streamlit
playwright==1.49.1
    # via
    #   crawl4ai
    #   tf-playwright-stealth
postgrest==0.19.3
    # via supabase
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via streamlit
psutil==6.1.1
    # via crawl4ai
pyarrow==19.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   anthropic
    #   cohere
    #   crawl4ai
    #   gotrue
    #   groq
    #   litellm
    #   mistralai
    #   ollama
    #   openai
    #   postgrest
    #   pydantic-ai-slim
    #   pydantic-graph
pydantic-ai==0.0.21
    # via rag-owu (pyproject.toml)
pydantic-ai-slim==0.0.21
    # via pydantic-ai
pydantic-core==2.27.2
    # via
    #   cohere
    #   pydantic
pydantic-graph==0.0.21
    # via pydantic-ai-slim
pydeck==0.9.1
    # via streamlit
pyee==12.0.0
    # via playwright
pygments==2.19.1
    # via rich
pyopenssl==25.0.0
    # via crawl4ai
python-dateutil==2.9.0.post0
    # via
    #   mistralai
    #   pandas
    #   realtime
    #   storage3
python-dotenv==1.0.1
    # via
    #   crawl4ai
    #   litellm
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via huggingface-hub
rank-bm25==0.2.2
    # via crawl4ai
realtime==2.3.0
    # via supabase
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   cohere
    #   crawl4ai
    #   huggingface-hub
    #   pydantic-ai-slim
    #   streamlit
    #   tiktoken
rich==13.9.4
    # via
    #   crawl4ai
    #   streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   groq
    #   httpx
    #   openai
snowballstemmer==2.2.0
    # via crawl4ai
soupsieve==2.6
    # via beautifulsoup4
storage3==0.11.3
    # via supabase
streamlit==1.41.1
    # via rag-owu (pyproject.toml)
strenum==0.4.15
    # via supafunc
supabase==2.12.0
    # via rag-owu (pyproject.toml)
supafunc==0.9.3
    # via supabase
tenacity==9.0.0
    # via streamlit
tf-playwright-stealth==1.1.1
    # via crawl4ai
tiktoken==0.8.0
    # via litellm
tokenizers==0.21.0
    # via
    #   cohere
    #   litellm
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   openai
types-requests==2.32.0.20241016
    # via cohere
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   altair
    #   anthropic
    #   anyio
    #   cohere
    #   groq
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   pyopenssl
    #   realtime
    #   referencing
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via mistralai
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
watchdog==6.0.0
    # via streamlit
websockets==14.2
    # via realtime
xxhash==3.5.0
    # via crawl4ai
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
